+++ database/migrations/2024_11_18_234403_sms_misr_settings.php
@@ -1,0 +1,15 @@
+<?php
+
+use Spatie\LaravelSettings\Migrations\SettingsMigration;
+
+return new class extends SettingsMigration
+{
+    public function up(): void
+    {
+        $this->migrator->add('sms-misr.sms_misr_username', '');
+        $this->migrator->add('sms-misr.sms_misr_password', '');
+        $this->migrator->add('sms-misr.sms_misr_sender', '');
+        $this->migrator->add('sms-misr.sms_misr_environment', 'live');
+        $this->migrator->add('sms-misr.sms_misr_active', true);
+    }
+};