+++ tests/database/factories/UserFactory.php
@@ -17,6 +17,7 @@ class UserFactory extends Factory
             'email' => $this->faker->unique()->safeEmail(),
             'email_verified_at' => now(),
             'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
+            'phone' => '+201000000000',
             'remember_token' => Str::random(10),
         ];
     }